+++ sdk/storage/Azure.Storage.Common/tests/CommonTestBase.cs
@@ -29,8 +29,9 @@ namespace Azure.Storage.Test
     BlobClientOptions.ServiceVersion.V2023_01_03,
     BlobClientOptions.ServiceVersion.V2023_05_03,
     BlobClientOptions.ServiceVersion.V2023_08_03,
-    RecordingServiceVersion = BlobClientOptions.ServiceVersion.V2023_08_03,
-    LiveServiceVersions = new object[] { BlobClientOptions.ServiceVersion.V2023_01_03 })]
+    BlobClientOptions.ServiceVersion.V2023_11_03,
+    RecordingServiceVersion = BlobClientOptions.ServiceVersion.V2023_11_03,
+    LiveServiceVersions = new object[] { BlobClientOptions.ServiceVersion.V2023_08_03 })]
     public abstract class CommonTestBase : StorageTestBase<StorageTestEnvironment>
     {
         protected readonly BlobClientOptions.ServiceVersion _serviceVersion;